move RcraProfileSerializer to core app
