+++ server/apps/sites/serializers/__init__.py
@@ -2,7 +2,6 @@ from .address_ser import AddressSerializer
 from .contact_ser import ContactSerializer, RcraPhoneSerializer
 from .profile_ser import (
     RcraPermissionSerializer,
-    RcraProfileSerializer,
     RcraSitePermissionSerializer,
 )
 from .site_ser import RcraSiteSerializer, SiteSerializer
