+++ server/apps/sites/serializers/profile_ser.py
@@ -1,10 +1,8 @@
 from rest_framework import serializers
 from rest_framework.exceptions import APIException
-from rest_framework.serializers import ModelSerializer
 
-from apps.sites.models.profile_models import RcraSitePermission
+from apps.sites.models import RcraSitePermission
 
-from ...core.models import RcraProfile
 from .base_ser import SitesBaseSerializer
 from .site_ser import SiteSerializer
 
@@ -163,51 +161,4 @@ class RcraPermissionSerializer(RcraSitePermissionSerializer):
             "eManifest",
             "WIETS",
             "myRCRAid",
-        ]
-
-
-class RcraProfileSerializer(ModelSerializer):
-    """
-    Model serializer for marshalling/unmarshalling a user's RcraProfile
-    """
-
-    user = serializers.StringRelatedField()
-    rcraSites = RcraSitePermissionSerializer(
-        source="permissions",
-        required=False,
-        many=True,
-    )
-    phoneNumber = serializers.CharField(
-        source="phone_number",
-        required=False,
-    )
-    rcraAPIID = serializers.CharField(
-        source="rcra_api_id",
-        required=False,
-    )
-    rcraAPIKey = serializers.CharField(
-        source="rcra_api_key",
-        required=False,
-        write_only=True,
-    )
-    rcraUsername = serializers.CharField(
-        source="rcra_username",
-        required=False,
-    )
-    apiUser = serializers.BooleanField(
-        source="is_api_user",
-        required=False,
-        allow_null=False,
-    )
-
-    class Meta:
-        model = RcraProfile
-        fields = [
-            "user",
-            "rcraAPIID",
-            "rcraAPIKey",
-            "rcraUsername",
-            "rcraSites",
-            "phoneNumber",
-            "apiUser",
         ]
