+++ server/apps/sites/tests/test_serializers.py
@@ -1,5 +1,3 @@
-import json
-
 import pytest
 
 from apps.sites.models import RcraSitePermission
